Avoid testing torchscript on deprecated functions

+++ test/torchaudio_unittest/prototype/functional/librosa_compatibility_test_impl.py
@@ -56,7 +56,7 @@ class Functional(TestBaseMixin):
             octwidth=octwidth,
             norm=norm,
             base_c=True,
-            dtype=float,
+            dtype=np.float32,
         )
 
         self.assertEqual(torchaudio_fbank, librosa_fbank.T)
